move stuff out of mod.rs and into a root layout container
